+++ scrapper/src/main/java/edu/java/scrapper/domain/model/chat/Chat.java
@@ -20,7 +20,7 @@ public class Chat {
     @Id
     long id;
 
-    int state;
+    short state;
 
     public Chat(long id) {
         this.id = id;
