+++ scrapper/src/main/java/edu/java/scrapper/service/processing_services/TgChatService.java
@@ -12,5 +12,5 @@ public interface TgChatService {
 
     Optional<Chat> getChat(long tgChatId);
 
-    void setState(long chatId, int state) throws NotExistException;
+    void setState(long chatId, short state) throws NotExistException;
 }
